refac: route style
