+++ test-web/src/routes/style/dataeditstyle.js
@@ -1,0 +1,17 @@
+const navy = '#0F3659';
+
+const style = {
+  fixed: {
+    marginTop: '80px',
+    right: '0',
+    left: '80px',
+    zIndex: 1,
+    width: 'fit-content',
+    borderRadius: '0',
+    display: 'flex',
+    justifyContent: 'space-between',
+    height: '70px',
+  },
+};
+
+export default style;
